login/requirements.txt
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
Flask==2.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
PyJWT==1.4.2
MarkupSafe==2.1.1
pycparser==2.21
pymongo==4.0.1
Werkzeug==2.1.2
zipp==3.8.0